refactor: rename core api to services
