+++ aiarena/core/management/commands/cancelmatches.py
@@ -1,6 +1,6 @@
 from django.core.management.base import BaseCommand, CommandError
 
-from aiarena.core.api import Matches
+from aiarena.core.services import Matches
 from aiarena.core.models import Match
