+++ aiarena/core/management/commands/generatestats.py
@@ -3,7 +3,7 @@ from django.db import transaction
 
 from django_pglocks import advisory_lock
 
-from aiarena.core.api.bot_statistics import BotStatistics
+from aiarena.core.services.bot_statistics import BotStatistics
 from aiarena.core.models import Competition, CompetitionParticipation
