+++ aiarena/core/management/commands/timeoutovertimematches.py
@@ -1,7 +1,7 @@
 from django.core.management.base import BaseCommand
 from django.db import transaction
 
-from aiarena.core.api import Matches
+from aiarena.core.services import Matches
 
 
 class Command(BaseCommand):
