+++ aiarena/core/services/bot_statistics.py
@@ -3,7 +3,7 @@ from django.db.models import Max
 
 from django_pglocks import advisory_lock
 
-from aiarena.core.api.internal.statistics.elo_graphs_generator import EloGraphsGenerator
+from aiarena.core.services.internal.statistics.elo_graphs_generator import EloGraphsGenerator
 from aiarena.core.models import CompetitionParticipation, Map, Match, MatchParticipation, Result
 from aiarena.core.models.competition_bot_map_stats import CompetitionBotMapStats
 from aiarena.core.models.competition_bot_matchup_stats import CompetitionBotMatchupStats
